Centered better
Ok so tile width is a bit weird. 


I added the constant "weirdPadding" awhile ago, when an item is added I typically do x += tileSize - weirdPadding. This makes the grid line up better. 

Weirdpadding needs to be instanced in the above node, gameWindow, so it can be used when calculating the space the grid takes, so grid padding can be calculated.

AHH